+++ tasks.rb
@@ -3088,6 +3088,9 @@ if ARGV[0].to_sym == :create_release_zip
   if ENV['CI']
     # CI doesn't have git, so default to everything
     git_files = Dir['**/*.*']
+    git_files -= Dir['workflow/tests/run*/*.*']
+    git_files -= Dir['workflow/tests/test_results/*.*']
+    git_files -= Dir['workflow/tests/test_files/**/*.*']
   else
     # Only include files under git version control
     command = 'git ls-files'
@@ -3114,30 +3117,7 @@ if ARGV[0].to_sym == :create_release_zip
            'workflow/sample_files/*.*',
            'workflow/tests/*.rb',
            'workflow/tests/**/*.csv',
-           'workflow/tests/**/*.xml',
-           'documentation/index.html',
-           'documentation/_static/**/*.*']
-
-  if not ENV['CI']
-    # Generate documentation
-    puts 'Generating documentation...'
-    command = 'sphinx-build -b singlehtml docs/source documentation'
-    begin
-      `#{command}`
-      if not File.exist? File.join(File.dirname(__FILE__), 'documentation', 'index.html')
-        puts 'Documentation was not successfully generated. Aborting...'
-        exit!
-      end
-    rescue
-      puts "Command failed: '#{command}'. Perhaps sphinx needs to be installed?"
-      exit!
-    end
-
-    fonts_dir = File.join(File.dirname(__FILE__), 'documentation', '_static', 'fonts')
-    if Dir.exist? fonts_dir
-      FileUtils.rm_r(fonts_dir)
-    end
-  end
+           'workflow/tests/**/*.xml']
 
   # Create zip files
   require 'zip'
@@ -3147,23 +3127,14 @@ if ARGV[0].to_sym == :create_release_zip
   Zip::File.open(zip_path, create: true) do |zipfile|
     files.each do |f|
       Dir[f].each do |file|
-        if file.start_with? 'documentation'
-          # always include
-        else
-          if not git_files.include? file
-            next
-          end
+        if not git_files.include? file
+          next
         end
+
         zipfile.add(File.join('OpenStudio-ERI', file), file)
       end
     end
-    puts "Wrote file at #{zip_path}."
   end
-
-  # Cleanup
-  if not ENV['CI']
-    FileUtils.rm_r(File.join(File.dirname(__FILE__), 'documentation'))
-  end
-
+  puts "Wrote file at #{zip_path}."
   puts 'Done.'
 end