+++ Changelog.md
@@ -8,7 +8,8 @@ __New Features__
   - Compressor lockout temperatures for Rated Home defaults and Reference Home updated from 0F (single/two stage) and -20F (variable speed) to 5F and -10F, respectively.
   - Dual-fuel heat pumps now have the HP sized using the greater of the heating design temperature and 25F.
   - Pan heaters no longer operate when the outdoor temperature is below the compressor lockout temperature.
-- Adds a warning if there are more than 300 unmet hours for heating or cooling.
+- Adds a warning if there are more than 300 unmet hours for heating or cooling; setpoint tolerance increased from 0.36 deg-F (0.2 deg-C) to 0.5 deg-F.
+- Runtime speed improvements for homes with multiple ground source heat pumps.
 - Utility bill scenarios can now be defined for the Rated/Reference Homes; see [the documentation](https://openstudio-eri.readthedocs.io/en/latest/workflow_inputs.html#hpxml-utility-bill-scenarios) for more information.
 - The `Battery/NominalCapacity` input is no longer used.
 - Output updates: